+++ cf/constructlist.py
@@ -400,7 +400,7 @@ class ConstructList(list, Container, cfdm.Container):
         verbose=None,
         ignore_data_type=False,
         ignore_fill_value=False,
-        ignore_properties=(),
+        ignore_properties=None,
         ignore_compression=False,
         ignore_type=False,
         ignore=(),
@@ -457,11 +457,7 @@ class ConstructList(list, Container, cfdm.Container):
 
             {{verbose: `int` or `str` or `None`, optional}}
 
-            ignore_properties: sequence of `str`, optional
-                The names of properties of the field construct (not
-                the TODO metadata constructs) to omit from the
-                comparison. Note that the "Conventions" property is
-                always omitted by default.
+            {{ignore_properties: (sequence of) `str`, optional}}
 
             {{ignore_data_type: `bool`, optional}}