+++ cf/aggregate.py
@@ -177,9 +177,10 @@ class _Meta:
             relaxed_units: `bool`, optional
                 If True then assume that field and metadata constructs
                 with the same identity but missing units actually have
-                equivalent (but unspecified) units, so that aggregation
-                may occur. By default such field constructs are not
-                aggregatable.
+                equivalent (but unspecified) units, so that
+                aggregation may occur. Also assumes that invalid but
+                otherwise equal units are equal. By default such field
+                constructs are not aggregatable.
 
             allow_no_identity: `bool`, optional
                 If True then assume that field and metadata constructs
@@ -237,6 +238,7 @@ class _Meta:
         )
 
         self.relaxed_identities = relaxed_identities
+        self.relaxed_units = relaxed_units
         self.strict_identities = strict_identities
         self.field_identity = field_identity
         self.ncvar_identities = ncvar_identities
@@ -659,34 +661,14 @@ class _Meta:
         # ------------------------------------------------------------
         self.msr = {}
         info_msr = {}
-        copied_field = False
         for key, msr in f.cell_measures(todict=True).items():
-            # If the measure is an external variable, remove it because
-            # the dimensions are not known so there is no way to tell if the
-            # aggregation should have changed it. (This is sufficiently
-            # sensible behaviour for now, but will be reviewed in future.)
-            # Note: for CF <=1.8 only cell measures can be external variables.
-            if msr.nc_get_external():
-                # Only create one copy of field if there is >1 external measure
-                if not copied_field:
-                    self.field = self.field.copy()  # copy as will delete msr
-                    f = self.field
-                    copied_field = True
+            if not self.cell_measure_has_measure(msr):
+                return
 
-                f.del_construct(key)
-
-                if is_log_level_info(logger):
-                    logger.info(
-                        f"Removed {msr.identity()!r} construct from a copy "
-                        f"of input field {f.identity()!r} pre-aggregation "
-                        "because it is an external variable so it "
-                        "is not possible to determine the influence the "
-                        "aggregation process should have on it."
-                    )
-
-                continue
-
-            if not self.cell_measure_has_data_and_units(msr):
+            if (
+                not msr.nc_get_external()
+                and not self.cell_measure_has_data_and_units(msr)
+            ):
                 return
 
             # Find the canonical units for this cell measure
@@ -712,15 +694,30 @@ class _Meta:
             else:
                 info_msr[units] = []
 
-            info_msr[units].append({"key": key, "axes": axes})
+            # Store the external status
+            if msr.nc_get_external():
+                external = msr.nc_get_variable(None)
+            else:
+                external = None
+
+            info_msr[units].append(
+                {
+                    "measure": msr.get_measure(),
+                    "key": key,
+                    "axes": axes,
+                    "external": external,
+                }
+            )
 
         # For each cell measure's canonical units, sort the
         # information by axis identities.
         for units, value in info_msr.items():
             value.sort(key=itemgetter("axes"))
             self.msr[units] = {
+                "measure": tuple([v["measure"] for v in value]),
                 "keys": tuple([v["key"] for v in value]),
                 "axes": tuple([v["axes"] for v in value]),
+                "external": tuple([v["external"] for v in value]),
             }
 
         # ------------------------------------------------------------
@@ -858,21 +855,32 @@ class _Meta:
         _canonical_units = self._canonical_units
 
         if identity in _canonical_units:
-            if var_units:
+            if var_units.isvalid:
+                if var_units:
+                    for u in _canonical_units[identity]:
+                        if var_units.equivalent(u):
+                            return u
+
+                    # Still here?
+                    _canonical_units[identity].append(var_units)
+                elif relaxed_units or variable.dtype.kind in ("S", "U"):
+                    return _no_units
+
+            elif relaxed_units:
                 for u in _canonical_units[identity]:
-                    if var_units.equivalent(u):
+                    if u.isvalid:
+                        continue
+
+                    if var_units.__dict__ == u.__dict__:
                         return u
 
                 # Still here?
                 _canonical_units[identity].append(var_units)
-
-            elif relaxed_units or variable.dtype.kind in ("S", "U"):
-                var_units = _no_units
         else:
-            if var_units:
+            if var_units or (relaxed_units and not var_units.isvalid):
                 _canonical_units[identity] = [var_units]
             elif relaxed_units or variable.dtype.kind in ("S", "U"):
-                var_units = _no_units
+                return _no_units
 
         # Still here?
         return var_units
@@ -943,11 +951,29 @@ class _Meta:
         """
         if not msr.Units:
             self.message = f"{msr.identity()!r} cell measure has no units"
-            return
+            return False
 
         if not msr.has_data():
             self.message = f"{msr.identity()!r} cell measure has no data"
-            return
+            return False
+
+        return True
+
+    def cell_measure_has_measure(self, msr):
+        """True only if a cell measure has a measure.
+
+        :Parameters:
+
+            msr: `CellMeasure`
+
+        :Returns:
+
+            `bool`
+
+        """
+        if not msr.get_measure(False):
+            self.message = f"{msr.identity()!r} cell measure has no measure"
+            return False
 
         return True
 
@@ -1182,7 +1208,11 @@ class _Meta:
         # * whether or not there is a data array
         Type = f.construct_type
         Identity = self.identity
-        Units = self.units.formatted(definition=True)
+        if self.units.isvalid:
+            Units = self.units.formatted(definition=True)
+        else:
+            Units = self.units.units
+
         Cell_methods = self.cell_methods
         Data = self.has_data
         #        signature_append = signature.append
@@ -1269,8 +1299,10 @@ class _Meta:
         msr = self.msr
         x = [
             (
+                ("measure", msr[units]["measure"]),
                 ("units", units.formatted(definition=True)),
                 ("axes", msr[units]["axes"]),
+                ("external", msr[units]["external"]),
             )
             for units in sorted(msr)
         ]
@@ -1485,7 +1517,8 @@ def aggregate(
             If True then assume that field and metadata constructs
             with the same identity but missing units actually have
             equivalent (but unspecified) units, so that aggregation
-            may occur. By default such field constructs are not
+            may occur. Also assumes that invalid but otherwise equal
+            units are equal. By default such field constructs are not
             aggregatable.
 
         allow_no_identity: `bool`, optional
@@ -1867,7 +1900,7 @@ def aggregate(
 
             continue
 
-        if not meta[0].units.isvalid:
+        if not relaxed_units and not meta[0].units.isvalid:
             if is_log_level_info(logger):
                 x = ", ".join(set(repr(m.units) for m in meta))
                 logger.info(
@@ -2030,6 +2063,7 @@ def aggregate(
                         atol=atol,
                         verbose=verbose,
                         concatenate=concatenate,
+                        relaxed_units=relaxed_units,
                         copy=(copy or not exclude),
                     )
 
@@ -2929,7 +2963,14 @@ def _ok_coordinate_arrays(meta, axis, overlap, contiguous, verbose=None):
 
 @_manage_log_level_via_verbosity
 def _aggregate_2_fields(
-    m0, m1, rtol=None, atol=None, verbose=None, concatenate=True, copy=True
+    m0,
+    m1,
+    rtol=None,
+    atol=None,
+    verbose=None,
+    concatenate=True,
+    relaxed_units=False,
+    copy=True,
 ):
     """Aggregate two fields, returning the _Meta object of the
     aggregated field.
@@ -3116,6 +3157,7 @@ def _aggregate_2_fields(
             data = Data.concatenate(
                 (construct0.get_data(), construct1.get_data()),
                 axis,
+                relaxed_units=relaxed_units,
             )
             construct0.set_data(data, copy=False)
             if construct0.has_bounds():
@@ -3125,6 +3167,7 @@ def _aggregate_2_fields(
                         construct1.bounds.get_data(_fill_value=False),
                     ),
                     axis,
+                    relaxed_units=relaxed_units,
                 )
                 construct0.bounds.set_data(data, copy=False)
         else:
@@ -3135,6 +3178,7 @@ def _aggregate_2_fields(
                     construct0.get_data(_fill_value=False),
                 ),
                 axis,
+                relaxed_units=relaxed_units,
             )
             construct0.set_data(data)
             if construct0.has_bounds():
@@ -3144,6 +3188,7 @@ def _aggregate_2_fields(
                         construct0.bounds.get_data(_fill_value=False),
                     ),
                     axis,
+                    relaxed_units=relaxed_units,
                 )
                 construct0.bounds.set_data(data)
 
@@ -3194,12 +3239,16 @@ def _aggregate_2_fields(
         if direction0:
             # The fields are increasing along the aggregating axis
             data = Data.concatenate(
-                (parent0.get_data(), parent1.get_data()), axis
+                (parent0.get_data(), parent1.get_data()),
+                axis,
+                relaxed_units=relaxed_units,
             )
         else:
             # The fields are decreasing along the aggregating axis
             data = Data.concatenate(
-                (parent1.get_data(), parent0.get_data()), axis
+                (parent1.get_data(), parent0.get_data()),
+                axis,
+                relaxed_units=relaxed_units,
             )
 
         # Update the size of the aggregating axis in parent0