+++ Changelog.rst
@@ -8,6 +8,8 @@ version 3.14.1
 * New keyword parameter to `cf.unique_constructs`:
   ``ignore_properties``
   (https://github.com/NCAS-CMS/cf-python/issues/597)
+* Allow `cf.aggregate` and to retain external cell meausures
+  (https://github.com/NCAS-CMS/cf-python/issues/605)
 * Fixed bug in `cf.Field.collapse` that always retained invalid
   vertical coordinate reference systems
   (https://github.com/NCAS-CMS/cf-python/issues/599)